More QEMU logic
git-svn-id: svn://svn.code.sf.net/p/nuttx/code/trunk@3337 42af7a65-404d-4744-a932-0658087f49c3

+++ arch/x86/include/i486/io.h
@@ -113,7 +113,7 @@ static inline void outl(uint32_t regval, uint16_t port)
     );
 }
 
-static inline uint32_t inl(uint32_t port)
+static inline uint32_t inl(uint16_t port)
 {
   uint32_t regval;
   asm volatile(
